MultiServer, customserver: cache: rename module
